PM4_code/requirements.txt
pycryptodome
cryptography
numpy
Pillow
scipy
darkdetect